requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0

